moved cluster classes into miniClusterImpl
