+++ test/src/main/java/org/apache/accumulo/harness/AccumuloClusterHarness.java
@@ -26,11 +26,6 @@ import static org.junit.jupiter.api.Assumptions.assumeTrue;
 import java.io.IOException;
 import java.util.Properties;
 
-import org.apache.accumulo.cluster.AccumuloCluster;
-import org.apache.accumulo.cluster.ClusterControl;
-import org.apache.accumulo.cluster.ClusterUser;
-import org.apache.accumulo.cluster.ClusterUsers;
-import org.apache.accumulo.cluster.standalone.StandaloneAccumuloCluster;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
 import org.apache.accumulo.core.client.admin.SecurityOperations;
@@ -42,6 +37,11 @@ import org.apache.accumulo.harness.conf.AccumuloClusterConfiguration;
 import org.apache.accumulo.harness.conf.AccumuloClusterPropertyConfiguration;
 import org.apache.accumulo.harness.conf.StandaloneAccumuloClusterConfiguration;
 import org.apache.accumulo.miniclusterImpl.MiniAccumuloConfigImpl;
+import org.apache.accumulo.miniclusterImpl.cluster.AccumuloCluster;
+import org.apache.accumulo.miniclusterImpl.cluster.ClusterControl;
+import org.apache.accumulo.miniclusterImpl.cluster.ClusterUser;
+import org.apache.accumulo.miniclusterImpl.cluster.ClusterUsers;
+import org.apache.accumulo.miniclusterImpl.cluster.standalone.StandaloneAccumuloCluster;
 import org.apache.accumulo.server.ServerContext;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
